file location refactors for making it easy to find everything
