+++ src/main/java/main/java/Repository/UserRepository.java
@@ -1,5 +1,6 @@
 package main.java.Repository;
 
+import main.java.Repository.Interfaces.IUserRepository;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
